ADX-271 add fullpage bulk file uploader
